requirements.txt
fastapi
uvicorn
sqlalchemy
passlib[bcrypt]
pyjwt
python-dotenv
python-multipart